final_project/requirements.txt
flask
sqlalchemy
flask_sqlalchemy
scikit-learn
numpy
pillow
werkzeug
